+Ahora no se puede eliminar Alumnos con PSs relacionadas, y se muestra mensaje de Error. +Solucionado errores en editar PS en caso de que la organización/área son nulos. +Cuando se crea una ps ahora el nº de disposición se muestra deshabilitado. +En las diferentes vistas en caso de que el nºdisposición/Área/Organización/Tutor/TituloProyecto/TipoPS sean nulos se muestra "- - -".

+++ WebApplication1/Controllers/AlumnosController.cs
@@ -112,6 +112,12 @@ namespace WebApplication1.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Alumno alumno = db.Alumnos.Find(id);
+
+            if (alumno.PSs.Count() != 0)
+            {
+                ViewBag.erroralumnops = "Acción no permitida!. Alumno con PSs relacionadas.";
+                return View(alumno);
+            }
             db.Alumnos.Remove(alumno);
             db.SaveChanges();
             return RedirectToAction("Index");
